fix: add loader to scenes, update demo
