+++ components/content/array-cameras/index.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { computed } from 'vue'
-import { TresCanvas, useRenderLoop } from '@tresjs/core'
-import { useAnimations, useGLTF, useTweakPane } from '@tresjs/cientos'
+import { TresCanvas } from '@tresjs/core'
+import { useProgress } from '@tresjs/cientos'
 import { PerspectiveCamera, Vector4, Vector3 } from 'three'
 import { useWindowSize } from '@vueuse/core'
 
@@ -50,54 +50,25 @@ cameraOptions.forEach(data => {
   cameras.push(currentCam)
 })
 
-const { scene: model, animations } = await useGLTF(
-  'https://raw.githubusercontent.com/Tresjs/assets/main/models/gltf/warcraft-3-alliance-footmanfanmade/source/Footman_RIG.glb',
-)
-const { actions, mixer } = useAnimations(animations, model)
-
-let currentAction = actions.Idle
-
-currentAction.play()
-
-const { pane } = useTweakPane()
-
-const animationList = pane.addBlade({
-  view: 'list',
-  label: 'Animations',
-  options: [
-    { text: 'Idle', value: 'Idle' },
-    { text: 'SwordAndShieldCrouchBlockIdle', value: 'SwordAndShieldCrouchBlockIdle' },
-    { text: 'SwordAndShieldDeath_2', value: 'SwordAndShieldDeath_2' },
-    { text: 'SwordAndShieldIdle', value: 'SwordAndShieldIdle' },
-    { text: 'SwordAndShieldKick', value: 'SwordAndShieldKick' },
-    { text: 'SwordAndShieldRun(back)', value: 'SwordAndShieldRun(back)' },
-    { text: 'SwordAndShieldRun', value: 'SwordAndShieldRun' },
-    { text: 'SwordAndShieldSlash_2', value: 'SwordAndShieldSlash_2' },
-    { text: 'SwordAndShieldSlash', value: 'SwordAndShieldSlash' },
-    { text: 'T-Pose', value: 'T-Pose' },
-  ],
-  value: 'Idle',
-})
-
-animationList.on('change', value => {
-  currentAction.stop()
-  currentAction = actions[value.value]
-  currentAction.play()
-})
-
-const { onLoop } = useRenderLoop()
-
-onLoop(() => {
-  if (mixer) {
-    mixer.update(0.01)
-  }
-})
+const { hasFinishLoading, progress, items } = await useProgress()
 </script>
 <template>
+  <Transition
+    name="fade-overlay"
+    enter-active-class="opacity-1 transition-opacity duration-200"
+    leave-active-class="opacity-0 transition-opacity duration-200"
+  >
+    <div
+      v-show="!hasFinishLoading"
+      class="absolute bg-grey-600 t-0 l-0 w-full h-full z-20 flex justify-center items-center text-black font-mono"
+    >
+      <div class="w-200px">Loading... {{ progress }} %</div>
+    </div>
+  </Transition>
   <TresCanvas window-size clear-color="#82DBC5" class="over-hidden">
     <TresArrayCamera :args="[cameras]" :position="[0, 2, 5]" />
     <Suspense>
-      <primitive :object="model" />
+      <model />
     </Suspense>
     <TresAmbientLight :color="0xffffff" :intensity="1" />
     <TresSpotLight :color="0xffffff" :intensity="100" :position="[0, 0, 5]" />
